# Documentation-financeAPI-1.0.0-swagger.yaml
openapi: 3.0.0
info:
  description: API documentation for Finance Management System
  version: "1.0.0"
  title: Finance API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/FIKRANELYAFIT/financeAPI/1.0.0
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token to access these API endpoints
  schemas:
...
    Penjualan:
      type: object
      properties:
        id:
          type: integer
          example: 1
        total_penjualan:
          type: number
          example: 100000
        tanggal:
          type: string
          format: date
          example: '2023-12-01'
    PenjualanInput:
      type: object
      required:
        - total_penjualan
        - tanggal
      properties:
        total_penjualan:
          type: number
          example: 100000
        tanggal:
          type: string
          format: date
          example: '2023-12-01'

    # Pengeluaran Gaji Schemas
...
    Rugi:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nama:
          type: string
          example: Kerusakan Mesin
        jumlah:
          type: number
          example: 1500000
        tanggal:
          type: string
          format: date
          example: '2023-12-10'
    RugiInput:
      type: object
      required:
        - nama
        - jumlah
        - tanggal
      properties:
        nama:
          type: string
          example: Kerusakan Mesin
        jumlah:
          type: number
          example: 1500000
        tanggal:
          type: string
          format: date
          example: '2023-12-10'

    # Total Penjualan Schemas
...
paths:
...
  /penjualans:
    get:
      tags:
        - Penjualans
      summary: Retrieve all penjualans
      operationId: getAllPenjualans
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of penjualans.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Penjualan'
        '500':
          description: Server error

    post:
      tags:
        - Penjualans
      summary: Create a new penjualan
      operationId: createPenjualan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PenjualanInput'
      responses:
        '201':
          description: Penjualan created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Penjualan'
        '400':
          description: Invalid input
        '500':
          description: Server error
  /penjualans/{id}:
    get:
      tags:
        - Penjualans
      summary: Retrieve a penjualan by ID
      operationId: getPenjualanById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Penjualan ID
          schema:
            type: integer
      responses:
        '200':
          description: A single penjualan object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Penjualan'
        '404':
          description: Penjualan not found
        '500':
          description: Server error

    put:
      tags:
        - Penjualans
      summary: Update a penjualan by ID
      operationId: updatePenjualanById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Penjualan ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PenjualanInput'
      responses:
        '200':
          description: Penjualan updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Penjualan updated successfully
        '400':
          description: Invalid input
        '404':
          description: Penjualan not found
        '500':
          description: Server error

    delete:
      tags:
        - Penjualans
      summary: Delete a penjualan by ID
      operationId: deletePenjualanById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Penjualan ID
          schema:
            type: integer
      responses:
        '200':
          description: Penjualan deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Penjualan deleted successfully
        '404':
          description: Penjualan not found
        '500':
          description: Server error
...
  /rugis:
    get:
      tags:
        - Rugis
      summary: Retrieve all rugis
      operationId: getAllRugis
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of rugis.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rugi'
        '500':
          description: Server error

    post:
      tags:
        - Rugis
      summary: Create a new rugi
      operationId: createRugi
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RugiInput'
      responses:
        '201':
          description: Rugi created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rugi'
        '400':
          description: Invalid input
        '500':
          description: Server error
  /rugis/{id}:
    get:
      tags:
        - Rugis
      summary: Retrieve a rugi by ID
      operationId: getRugiById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Rugi ID
          schema:
            type: integer
      responses:
        '200':
          description: A single rugi object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rugi'
        '404':
          description: Rugi not found
        '500':
          description: Server error

    put:
      tags:
        - Rugis
      summary: Update a rugi by ID
      operationId: updateRugiById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Rugi ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RugiInput'
      responses:
        '200':
          description: Rugi updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Rugi updated successfully
        '404':
          description: Rugi not found
        '500':
          description: Server error

    delete:
      tags:
        - Rugis
      summary: Delete a rugi by ID
      operationId: deleteRugiById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Rugi ID
          schema:
            type: integer
      responses:
        '200':
          description: Rugi deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Rugi deleted successfully
        '404':
          description: Rugi not found
        '500':
          description: Server error
...